Dijkstra: Shortest Reach 2
Dijkstra: Shortest Reach 2

https://www.hackerrank.com/challenges/dijkstrashortreach/problem

Given an undirected graph and a starting node, determine the lengths of the shortest paths from the starting node to all other nodes in the graph. If a node is unreachable, its distance is -1. Nodes will be numbered consecutively from  to , and edges will have varying distances or lengths.